refactor: internal imports rewritten
