+++ lib/components/AutoSuggest/AutoSuggest.tsx
@@ -20,17 +20,17 @@ import React, {
 	useState,
 } from 'react';
 
-import { useMedia } from '../../hooks/useMedia';
+import { useMedia } from '../../hooks/useMedia/useMedia';
 import { useWindowScrollLock } from '../../hooks/useWindowScrollLock';
 import { setRef, useId } from '../../utils';
-import { Box } from '../Box';
-import { Button } from '../Button';
-import { Icon } from '../Icon';
-import { Portal } from '../Portal';
-import { Positioner } from '../Positioner';
+import { Box } from '../Box/Box';
+import { Button } from '../Button/Button';
+import { Icon } from '../Icon/Icon';
+import { Portal } from '../Portal/Portal';
+import { Positioner } from '../Positioner/Positioner';
 import { EAlignment } from '../Positioner/alignment';
-import { Text } from '../Text';
-import { TextInput } from '../TextInput';
+import { Text } from '../Text/Text';
+import { TextInput } from '../TextInput/TextInput';
 import { withEnhancedInput } from '../private/InputBase';
 
 import * as styles from './AutoSuggest.css';
