+++ lib/components/Button/Button.tsx
@@ -16,10 +16,11 @@ import React, {
 
 import type { TextFontWeight, TextSizeScale } from '../../themes';
 import { dataAttrs } from '../../utils/dataAttrs';
-import { Box, BoxLikeProps, useBox, type UseBoxProps } from '../Box';
-import { Icon } from '../Icon';
-import { ProgressSpinner } from '../ProgressSpinner';
-import { useTextStyles } from '../Text';
+import { Box } from '../Box/Box';
+import { BoxLikeProps, useBox, type UseBoxProps } from '../Box/useBox';
+import { Icon } from '../Icon/Icon';
+import { ProgressSpinner } from '../ProgressSpinner/ProgressSpinner';
+import { useTextStyles } from '../Text/useTextStyles';
 
 import * as styles from './Button.css';
 import type { ButtonSize, StyledButtonProps } from './Button.css';
