+++ lib/components/CheckBox/CheckBox.tsx
@@ -4,9 +4,9 @@ import React, { forwardRef, ReactNode, useEffect, useRef } from 'react';
 
 import { mergeRefs, noop } from '../../utils';
 import { dataAttrs } from '../../utils/dataAttrs';
-import { Box } from '../Box';
-import { Icon } from '../Icon';
-import { CheckableBase } from '../private/CheckableBase';
+import { Box } from '../Box/Box';
+import { Icon } from '../Icon/Icon';
+import { CheckableBase } from '../private/CheckableBase/CheckableBase';
 import { checkableIndicator } from '../private/CheckableBase/CheckableBase.css';
 
 import * as styles from './CheckBox.css';
