+++ lib/components/ColourInput/ColourInput.tsx
@@ -2,8 +2,9 @@ import { warning } from '@autoguru/utilities';
 import clsx from 'clsx';
 import * as React from 'react';
 
-import { Box, boxStyles } from '../Box';
-import { Text } from '../Text';
+import { Box } from '../Box/Box';
+import { boxStyles } from '../Box/boxStyles';
+import { Text } from '../Text/Text';
 import { withEnhancedInput } from '../private/InputBase';
 
 import * as styles from './ColourInput.css';
