+++ lib/components/FillHeightBox/FillHeightBox.tsx
@@ -4,7 +4,7 @@ import { ComponentProps, FunctionComponent } from 'react';
 import useWindowHeightFill, {
 	UseWindowHeightFillProps,
 } from '../../hooks/useWindowHeightFill/useWindowHeightFill';
-import { ScrollPane } from '../ScrollPane';
+import { ScrollPane } from '../ScrollPane/ScrollPane';
 
 type FillHeightBoxProps = Omit<UseWindowHeightFillProps, 'containerRef'> &
 	ComponentProps<typeof ScrollPane>;
