+++ lib/components/Flyout/Flyout.tsx
@@ -1,8 +1,8 @@
 import type { ComponentProps } from 'react';
 import * as React from 'react';
 
-import { Box } from '../Box';
-import { Positioner } from '../Positioner';
+import { Box } from '../Box/Box';
+import { Positioner } from '../Positioner/Positioner';
 
 export const Flyout = ({
 	children,
