+++ lib/components/SimplePagination/SimplePagination.tsx
@@ -3,9 +3,9 @@ import * as React from 'react';
 import { FunctionComponent } from 'react';
 
 import { noop } from '../../utils';
-import { Button } from '../Button';
-import { Icon } from '../Icon';
-import { Inline } from '../Inline';
+import { Button } from '../Button/Button';
+import { Icon } from '../Icon/Icon';
+import { Inline } from '../Inline/Inline';
 
 export enum EChangeDirection {
 	Previous = 'previous',
